+++ R/define_S4_class_retro.R
@@ -5,7 +5,7 @@
 #' @description Perform a retrospective analysis, successive removals of most recent years of data to evaluate resulting
 #' parameter estimates.
 #'
-#' @param x An S4 object of class \linkS4class{Assessment}.
+#' @param x An S4 object of class \linkS4class{Assessment} of \linkS4class{SRA}.
 #' @param nyr The maximum number of years to remove for the retrospective analysis.
 #' @param figure Indicates whether plots will be drawn.
 #' @param ... More arguments.
@@ -46,6 +46,24 @@ setMethod("retrospective", signature(x = "Assessment"),
             return(res)
           })
 
+
+#' @rdname retrospective
+#' @aliases retrospective,SRA-method
+#' @exportMethod retrospective
+setMethod("retrospective", signature(x = "SRA"),
+          function(x, nyr = 5, figure = TRUE) {
+            if(figure) {
+              old.warning <- options()$warn
+              options(warn = -1)
+              on.exit(options(warn = old.warning))
+
+              old_par <- par(no.readonly = TRUE)
+              on.exit(par(old_par), add = TRUE)
+            }
+            res <- SRA_retro(x, nyr)
+            if(figure) plot(res)
+            return(res)
+          })
 
 #' Class-\code{retro}
 #'